Done with the custom_broadcasting. Since from sdk 26 by registering to the manifest file does not work it was needed to be done dynamically. Moreover, while trying to go on to the third activity from second, as soon as broadcasting happens from the second activity the third activity cannot receive the broadcast cause it is not active when broadcast happens. So ass soon as we land on the third activity the data gets lost from broadcast. but when pressed back key and then again clicking onto the button works fine as this time the activity was on service but paused to data could be received. This issue could be solved by starting the service of launching the third activity on background before manually going to the third activity but from above sdk 26 this method was restricted. Not only that but also, on TextView the result couldn't be updated despite of being able to show the message on Toast. This was most probably because the TextView run on a different thread initiated by the onCreate() [not sure about this point].
Finally came to a conclusion of passing the message from 2nd to 3rd activity over intent.putExtra which is normal data passing the receiving the data from the 3rd activity the message was broadcast from there and the 3rd activity was manually closed using finish().
Then as soon as the 2nd activity returned the button and editText's Views were hidden and by receiving the broadCast message the output was set into the TextView on the 2nd activity.
This time it worked because the 2nd activity is active already.